Buncha Updates
Implementation of new kind of Observability by the Attacker in Events.
Implementation of new kind of Observability by the Attacker in File I/O and Generation, as well as regular Observability as well for Modal Specifications.
Removed Union operation as it shouldn't actually exist.
Manipulation of UI elements to respect the aforementioned changes.
Eh?