Windows Machines can now use the Program!
Outdated Graphviz code assumed that a Windows operating system would describe itself as 'Windows', when in fact they now attach their OS #.  'Windows10' is not 'Windows', so when searching for the requisite file paths, it did not find the matching variable, causing crashes. Now, if the OS includes the word "Windows", it will assign the entire phrase to just "Windows" and remove the numbers. It works.